% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostWin.R
\name{hostWin}
\alias{hostWin}
\title{hostWin}
\usage{
hostWin(
  appDir = getwd(),
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  workerId = "",
  quiet = FALSE,
  display.mode = c("auto", "normal", "showcase"),
  test.mode = getOption("shiny.testmode", FALSE)
)
}
\arguments{
\item{appDir}{see `host` parameter in `runApp()`}

\item{port}{see `port` parameter in `runApp()`}

\item{launch.browser}{see `launch.browser` parameter in `runApp()`}

\item{host}{see `host` parameter in `runApp()`}

\item{workerId}{see `workerId` parameter in `runApp()`}

\item{quiet}{see `quiet` parameter in `runApp()`}

\item{display.mode}{see `display.mode` parameter in `runApp()`}

\item{test.mode}{see `test.mode` parameter in `runApp()`}
}
\description{
hostWin
}
