% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_exe.R
\name{shiny.exe}
\alias{shiny.exe}
\title{shiny.exe}
\usage{
shiny.exe(
  appName,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"),
  appDir = getwd()
)
}
\arguments{
\item{appName}{the name of you shortcut or application}

\item{port}{see `port` parameter in `runApp()`}

\item{host}{see `host` parameter in `runApp()`}

\item{appDir}{see `host` parameter in `runApp()`}
}
\description{
shiny.exe
}
